move restricted_methods check to cli.rs
